requirements.txt
requests
streamlink
python-dotenv
